core: constants, token: try to initialize token mapping from url
